FIX UnicodeDecodeError in speed plugin
When dd print error message it use pretty UTF8 quotes, if locale
settings allow UTF8 output. So we got this error message this pretty
quotes and try to log it in our log message - we got UnicodeDecodeError.

It can be fixed at least in 2 way - we can "convert" output from remote
command in utf8, or we can force them to use only ASCII symbols. First
way - convert output to utf8 is more dangerous, because we have no
guarantee that we alway will receive correct utf8 stream. So as for now,
we can force "C" locale and leave everything else as is.

Change-Id: I3ac3f74995a419232c3806e9e782b88ae8d285c4
Closes-Bug: MCV-862